Lower the timeouts.
If it's taking more than 5 seconds for an operation, we have other
problems.
Also if it takes more than 10 seconds overall, consider the drive gone
so we fail and can restart the connection.  At the very least, it's
overloaded if it's taking this long.

We should be throttling to prevent this, not hanging out this long
waiting for responses.